ranking: set latest rank for newly appeared user

+++ modules/jb-module-ranking/src/main/java/ru/caramel/juniperbot/module/ranking/model/RankingInfo.java
@@ -45,7 +45,7 @@ public class RankingInfo {
 
     private long totalExp;
 
-    private int rank;
+    private long rank;
 
     private long cookies;
